Fix sort_order integer overflow
sort_order was declared as a Postgres integer (max ~2.1B) but tasks/lists
are seeded with Date.now() (~1.7 trillion ms), which overflowed on every
insert. Every task write failed with a 400 and was silently dropped by the
outbox's "non-network errors don't retry forever" rule — so the app looked
like it worked (optimistic cache) while nothing reached the database.
Widened the column to bigint.

+++ supabase/migrations/0001_init.sql
@@ -8,7 +8,7 @@ create table if not exists lists (
   user_id uuid not null references auth.users(id) on delete cascade,
   name text not null,
   color text,
-  sort_order integer not null default 0,
+  sort_order bigint not null default 0,
   created_at timestamptz not null default now()
 );
 
@@ -23,7 +23,7 @@ create table if not exists tasks (
   due_on date,
   due_at time,
   completed_at timestamptz,
-  sort_order integer not null default 0,
+  sort_order bigint not null default 0,
   created_at timestamptz not null default now(),
   updated_at timestamptz not null default now()
 );
